scripts/requirements.txt
openpyxl
requests
Pillow
pyyaml
